requirements.txt
aiogram
aioredis
environs

gspread
gspread-asyncio

oauth2client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib